# stms-backend/openapi.yaml
openapi: 3.0.3
info:
  title: STMS - Smart Trainer Management System API
  version: 1.0.0
  description: |
    Authentication & User Management API for Sprint 1.
    Multi-tenant athletics management platform with club-scoped data isolation.
  contact:
    name: STMS Team
    email: dev@stms.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.stms.example.com/v1
    description: Production
  - url: https://api-staging.stms.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 signed JWT access token (15 min expiry).
        Include in Authorization header: `Bearer <access_token>`
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HttpOnly, Secure, SameSite=Strict refresh token cookie (7 days)
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors (400 only)
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        firebaseUid:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [system_admin, club_admin, coach, athlete]
        clubIds:
          type: array
          items:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        activeClubId:
          type: string
          nullable: true
          pattern: '^[0-9a-fA-F]{24}$'
        status:
          type: string
          enum: [active, invited, deactivated, deleted]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ClubResponse:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        name:
          type: string
        slug:
          type: string
        branding:
          type: object
          properties:
            logoUrl:
              type: string
              nullable: true
            primaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
            secondaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
        settings:
          type: object
          properties:
            timezone:
              type: string
              example: "Asia/Kolkata"
            attendanceMinPercent:
              type: integer
              minimum: 0
              maximum: 100
              default: 75
            workoutVerificationRequired:
              type: boolean
              default: false
            notificationDefaults:
              type: object
        createdBy:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateClubRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          maxLength: 50
          description: Auto-generated from name if not provided
        branding:
          type: object
          properties:
            primaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
              default: "#1E3A8A"  # Blue
            secondaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
              default: "#F59E0B"  # Gold
        settings:
          type: object
          properties:
            timezone:
              type: string
              default: "Asia/Kolkata"
    UpdateClubSettingsRequest:
      type: object
      properties:
        branding:
          type: object
          properties:
            logoUrl:
              type: string
            primaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
            secondaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
        settings:
          type: object
          properties:
            timezone:
              type: string
            attendanceMinPercent:
              type: integer
              minimum: 0
              maximum: 100
            workoutVerificationRequired:
              type: boolean
            notificationDefaults:
              type: object
    ClubMemberResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        membership:
          type: object
          properties:
            role:
              type: string
              enum: [member, captain, alumni, head_coach, assistant_coach, specialist_coach]
            status:
              type: string
              enum: [active, pending, transferred_out]
            joinedAt:
              type: string
              format: date-time
            invitedBy:
              type: string
              nullable: true
  parameters:
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
paths:
...
  /clubs:
    post:
      tags: [Clubs]
      summary: Create new club (system_admin only)
      description: Creates club, default settings, and club_admin membership for creator
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClubRequest'
      responses:
        '201':
          description: Club created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires system_admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clubs/{id}:
    get:
      tags: [Clubs]
      summary: Get club details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Club details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not member of this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Clubs]
      summary: Update club settings (club_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClubSettingsRequest'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Requires club_admin role in this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /clubs/{id}/members:
    get:
      tags: [Clubs]
      summary: List club members with membership details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: role
          in: query
          schema:
            type: string
            enum: [member, captain, alumni, head_coach, assistant_coach, specialist_coach]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, transferred_out]
      responses:
        '200':
          description: Paginated member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClubMemberResponse'
                  meta:
                    $ref: '#/components/schemas/PaginatedResponse/properties/meta'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not member of this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== HEALTH CHECK ====================
...